slave: Use VhostUserProtocolFeatures more
Make the protcol_features member, and the result of
get_protocol_features be a VhostUserProtocolFeatures rather than a u64,
this stops someone implementing it returning one of the wrong feature
flags.

Note we can't do this to set_protocol_features because the bitflags!
can't hold the possibly other bits that might be set that could be
passed from the master;  we might be able to convert
acked_protocol_features if we're happy for that to be missing
unknown bits.

Signed-off-by: Dr. David Alan Gilbert <dgilbert@redhat.com>