initial support for handling slave requests
Not all slave requests are implemented, only the subset of messages used by
virtio-fs.

The SlaveReqHandler struct creates the slave-master socketpair, and handles
all incoming / outgoing message validation and handler routing.
SlaveReqHandler::new accepts a struct that implements VhostUserSlaveReqHandler,
overriding the default handlers for relevant slave requests.
SlaveReqHandler::new also returns a UnixStream that must be sent to the slave
via Master::set_slave_request_fd.

As both Master and Slave requests share the same wire-format, it could be easy
to mix up the two when handling requests, so this commit introduces PhantomData
markers to Endpoints and VhostUserMsgHeader that specify if they operate on
MasterReqs or SlaveReqs.

Lastly, some library internals, notably the Listener and Endpoint structs, have
been marked as pub(crate). This necessitates a small refactor of how Slaves
are constructed, introducing a SlaveListener struct which returns a valid Slave
when the socket accepts a connection.